+++ route4me-csharp-sdk/Route4MeSDKTest/Examples/Geocoding/BulkRemoveContactsFromDb.cs
@@ -22,9 +22,9 @@ namespace Route4MeSDK.Examples
                                             addressBookParameters,
                                             out ResultResponse resultResponse);
 
-            int[] addressIDs = response.Results
+            long[] addressIDs = response.Results
                                         .Where(x => x.AddressId != null)
-                                        .Select(x => (int)x.AddressId)
+                                        .Select(x => (long)x.AddressId)
                                         .ToArray();
 
             var fastBulkRemoveContacts = new FastProcessing.FastBulkRemoveContacts(ActualApiKey)
